requirements.txt
python-dotenv
pandas
numpy
scipy
pyyaml
alpaca-trade-api
